pyproject.toml
[build-system]
requires = ["scikit-build-core >=0.4.3", "nanobind >=1.3.2", "madness"]
build-backend = "scikit_build_core.build"

[tool.scikit-build.cmake]
args =[ 
  "-DCMAKE_CXX_FLAGS=-O3 -DNDEBUG -fcolor-diagnostics",
  "-DCMAKE_COLOR_MAKEFILE=ON",
  "-DCMAKE_CXX_FLAGS=-w",  # or use -Wno-everything if using Clang
  "-DCMAKE_C_FLAGS=-w"     # optional for pure C files
]
#generator = "Ninja"


[project]
name = "frayedends"
version = "0.0.1"
description = "Python interface for Orbital Refinement"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@address.com" },
]
classifiers = [
    "License :: ???",
]
dependencies = [
    "numpy",
    "pyscf",
    "qulacs",
    "tequila-basic",
    "block2",
    "nanobind",
    "ruff",
    "clang-format>=17.0.6"
]

[project.urls]
Homepage = "https://github.com/..."

[tool.scikit-build]
# Protect the configuration against future changes in scikit-build-core
minimum-version = "0.4"

# Setuptools-style build caching in a local directory
build-dir = "build/{wheel_tag}"

# Build stable ABI wheels for CPython 3.12+
wheel.py-api = "cp312"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # E: pycodestyle errors
    "E",
    # I: import sorting
    "I"
]
ignore = [
    # E501: Line too long
    # Mostly taken care of by the formatter, remaining issues are string literals and comments which is not a priority
    # at the moment
    "E501",
    # E731: Do not assign a `lambda` expression, use a `def`
    # Annoying because it will split a one line lambda into a two line def, and then add empty lines before and after
    # the def, so now the code is four times as long.
    "E731",
    # E741/E742/E743: Ambiguous variable/class/function name
    # Not a significant issue when using a good font, and clashes with some conventions, e.g. I for identity
    "E741",
    "E742",
    "E743",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"